CalculateEstimatedStopsPerHour Green

+++ EMLib/EmployeeMetrics.cs
@@ -8,7 +8,7 @@
         public double TotalTimeDriving { get; set; }
         public double EstimatedTimeFromTerminal { get; set; }
         public int RouteNumber { get; set; }
-        public int StopsPerHour { get; set; }
+        public double StopsPerHour { get; set; }
         public int TotalStops { get; set; }
 
         public double MilesDriven { get; set; }
